+++ Tests/Unit/Controller/ProjectControllerTest.php
@@ -26,7 +26,10 @@ class ProjectControllerTest extends UnitTestCase
             'settings',
             [
                 'filter' => [],
-                'cache' => ['lifetime' => '']
+                'cache' => [
+                    'lifetime' => '',
+                    'inactive' => '0'
+                ]
             ]
         );
 
@@ -56,6 +59,54 @@ class ProjectControllerTest extends UnitTestCase
         $cache = $this->getMockBuilder(FrontendInterface::class)->disableOriginalConstructor()->getMock();
         $cache->expects(self::once())->method('get')->willReturn(false);
         $cache->expects(self::once())->method('set');
+        $projectController->_set('cache', $cache);
+
+        $projectController->listAction();
+    }
+
+    public function testListActionWithInactiveCache(): void
+    {
+        $projects = array_fill(0, 2, 'Project');
+
+        $projectController = $this->getAccessibleMock(ProjectController::class, null, [], '', false);
+
+        $projectController->_set(
+            'settings',
+            [
+                'filter' => [],
+                'cache' => [
+                    'lifetime' => '',
+                    'inactive' => '1'
+                ]
+            ]
+        );
+
+        $cachingService = $this->getMockBuilder(CachingService::class)->disableOriginalConstructor()->getMock();
+        $cachingService->expects(self::once())->method('calculateCacheIdentifier')->willReturn('does_not_matter');
+        $projectController->_set('cachingService', $cachingService);
+
+        $configurationManager = $this->getMockBuilder(ConfigurationManager::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+        $configurationManager->expects(self::once())->method('getContentObject');
+        $projectController->_set('configurationManager', $configurationManager);
+
+        $filterFactory = $this->getMockBuilder(FilterFactory::class)->disableOriginalConstructor()->getMock();
+        $filterFactory->expects(self::once())->method('makeProjectFilter')->willReturn(new ProjectFilter());
+        $projectController->_set('filterFactory', $filterFactory);
+
+        $projectRepository = $this->getMockBuilder(ProjectRepository::class)->disableOriginalConstructor()->getMock();
+        $projectRepository->expects(self::once())->method('findForFilter')->willReturn($projects);
+        $projectController->_set('projectRepository', $projectRepository);
+
+        $view = $this->getMockBuilder(ViewInterface::class)->getMock();
+        $view->expects(self::once())->method('assign')->with('projects', $projects);
+        $view->expects(self::once())->method('render')->willReturn(json_encode($projects, JSON_THROW_ON_ERROR));
+        $projectController->_set('view', $view);
+
+        $cache = $this->getMockBuilder(FrontendInterface::class)->disableOriginalConstructor()->getMock();
+        $cache->expects(self::never())->method('get');
+        $cache->expects(self::never())->method('set');
         $projectController->_set('cache', $cache);
 
         $projectController->listAction();